% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altSEEExploreModule.R
\name{altSEEExploreServer}
\alias{altSEEExploreServer}
\title{altSEE explore tab server}
\usage{
altSEEExploreServer(id = "explore", variables)
}
\description{
altSEE explore tab server
}
\keyword{internal}
