% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose.dir2}
\alias{choose.dir2}
\title{-e "POSIX path of (choose folder{{prompt}}{{default}})"'}
\usage{
choose.dir2(default = NA, caption = NA)
}
\description{
-e "POSIX path of (choose folder{{prompt}}{{default}})"'
}
