% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_exception}
\alias{add_exception}
\title{A function that will add an exception entry to the global exception data}
\usage{
add_exception(title, type, func_call, message, full_message, time)
}
\arguments{
\item{title}{\code{str} title of the exception}

\item{type}{\code{str} type of the exception c("warning", "error")}

\item{func_call}{\code{str} function call that caused the exception}

\item{message}{\code{str} message of the exception}

\item{full_message}{\code{str} full message of the exception}

\item{time}{\code{POSIXct} time of the exception}
}
\value{
does not return anything but adds an exception to the global exception data
}
\description{
A function that will add an exception entry to the global exception data
}
\keyword{internal}
