% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error_handler}
\alias{error_handler}
\title{An wrapper function to handle errors and warnings
Will create a notification and add the exception to the global exception data}
\usage{
error_handler(func, component_name, ...)
}
\arguments{
\item{func}{function that is wrapped}

\item{component_name}{\code{str} name of the component (will be reported in the exception message)}

\item{...}{arguments to be passed to the function}
}
\value{
Does not return anything but will create a notification and add the exception to the global exception data
}
\description{
An wrapper function to handle errors and warnings
Will create a notification and add the exception to the global exception data
}
\keyword{internal}
