% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDataPath}
\alias{getDataPath}
\title{Function to convert data paths if on windows}
\usage{
getDataPath(datapath)
}
\arguments{
\item{datapath}{string with path to file}
}
\value{
string
}
\description{
Function to convert data paths if on windows
}
\keyword{internal}
