% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceModule.R, R/reportModule.R
\name{inferenceServer}
\alias{inferenceServer}
\alias{reportServer}
\title{Server for inference tab}
\usage{
inferenceServer(id = "inference", variables, inputServerInput)

reportServer(
  id = "report",
  variables,
  inputServerInput,
  modelServerInput,
  inferenceServerInput
)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}

\item{inputServerInput}{list with reactive values containing the input of the inputServer module}

\item{modelServerInput}{list with reactive values containing the input of the modelServer module}

\item{inferenceServerInput}{list with reactive values containing the input of the inferenceServer module}
}
\value{
list of reactive inputs

list of reactive inputs
}
\description{
Server for inference tab

Server for report tab
}
\keyword{internal}
