% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceModule.R
\name{inferenceUI}
\alias{inferenceUI}
\title{inference UI}
\usage{
inferenceUI(id = "inference")
}
\value{
A shiny tagList object that contains the inference UI components
}
\description{
inference UI
}
\keyword{internal}
