% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModule.R, R/preprocessingModule.R,
%   R/qcModule.R
\name{importServer}
\alias{importServer}
\alias{preprocessingServer}
\alias{qcServer}
\title{Server for input tab}
\usage{
importServer(id = "import", variables)

preprocessingServer(id = "preprocessing", variables)

qcServer(id = "qc", variables)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}
}
\value{
list of reactive inputs

list of reactive inputs

list of reactive inputs
}
\description{
Server for input tab

Server for preprocessing tab

Server for input tab
}
\keyword{internal}
