% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModule.R, R/qcModule.R
\name{inputUI}
\alias{inputUI}
\title{input UI}
\usage{
inputUI(id = "import")

inputUI(id = "import")
}
\value{
A shiny tagList object that contains the input UI components

A shiny tagList object that contains the input UI components
}
\description{
input UI

input UI
}
\keyword{internal}
