% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeDetailPlots}
\alias{makeDetailPlots}
\title{Function for detailplot}
\usage{
makeDetailPlots(pe, clickInfo, detailServerInput, inputServerInput)
}
\arguments{
\item{pe}{QFeatures object}

\item{clickInfo}{clickInfo object with the selection of features in plot or DT data table}

\item{detailServerInput}{input object with reactive values passed through via the inference Server UI.}

\item{inputServerInput}{list with reactive values returned by inputModule Server}
}
\value{
ggplot object
}
\description{
Function for detailplot
}
\keyword{internal}
