% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeVolcanoPlot}
\alias{makeVolcanoPlot}
\title{Function for volcano plot}
\usage{
makeVolcanoPlot(dataset, clickInfo, input, ranges)
}
\arguments{
\item{dataset}{dataframe with variables logFC: log2 fold changes and pval: pvalues. Typically output from the hypothesisTest function from msqrob2}

\item{clickInfo}{clickInfo object with the selection of features in plot or DT data table}

\item{input}{input object with reactive values passed through via the server UI.}

\item{ranges}{ranges object from the selectio nof features in plot}
}
\value{
ggplot object
}
\description{
Function for volcano plot
}
\keyword{internal}
