% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelModule.R
\name{modelServer}
\alias{modelServer}
\title{Server for model tab}
\usage{
modelServer(id = "model", variables)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}
}
\value{
list of reactive inputs
}
\description{
Server for model tab
}
\keyword{internal}
