% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{page_assays_subset}
\alias{page_assays_subset}
\title{A function that will subset the assays of a QFeatures object}
\usage{
page_assays_subset(qfeatures, pattern = NULL)
}
\arguments{
\item{qfeatures}{\code{QFeatures} object to subset}

\item{pattern}{\code{str} pattern to match the assays names}
}
\value{
\code{QFeatures} object with the subsetted assays
}
\description{
A function that will subset the assays of a QFeatures object
}
\keyword{internal}
