% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotDensities}
\alias{plotDensities}
\alias{makeBoxplotFC}
\title{Function for plotting densities of features}
\usage{
plotDensities(pe, assayName, varName)

makeBoxplotFC(dataset, sel, regulation = "both")
}
\arguments{
\item{pe}{QFeatures object}

\item{assayName}{string with name of QFeatures assay}

\item{varName}{string with name of variable in colData used to color the density curves}

\item{dataset}{data frame returned by hypothesisTest function of msqrob2}

\item{sel}{selection of features to be included in plot}

\item{regulation}{string "up", "down" or "both" to plot only upregulated, downregulated or all DE features.}
}
\value{
ggplot object

baseplot object
}
\description{
Function for plotting densities of features

Function for plotting densities of features
}
\keyword{internal}
