% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotPCA}
\alias{plotPCA}
\title{Function for pca plot of samples}
\usage{
plotPCA(pe, assayName, varName)
}
\arguments{
\item{pe}{QFeatures object}

\item{assayName}{string with name of QFeatures assay}

\item{varName}{string with name of variable in colData used to color the density curves}
}
\value{
ggplot object
}
\description{
Function for pca plot of samples
}
\keyword{internal}
