% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingModule.R
\name{preprocessingUI}
\alias{preprocessingUI}
\title{preprocessing UI}
\usage{
preprocessingUI(id = "preprocessing")
}
\value{
A shiny tagList object that contains the preprocessing UI components
}
\description{
preprocessing UI
}
\keyword{internal}
