% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qfeatures_to_df}
\alias{qfeatures_to_df}
\title{Will convert a qfeatures object to a summary data.frame object}
\usage{
qfeatures_to_df(qfeatures)
}
\arguments{
\item{qfeatures}{a qfeatures object}
}
\value{
a data.frame object
}
\description{
Will convert a qfeatures object to a summary data.frame object
}
\keyword{internal}
