% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCombinedModule.R
\name{reportCombinedServer}
\alias{reportCombinedServer}
\title{Combined report server}
\usage{
reportCombinedServer(
  id = "report",
  variables,
  importServerInput,
  preprocessingServerInput,
  qcServerInput,
  modelServerInput,
  inferenceServerInput
)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values}

\item{importServerInput}{reactives from importServer}

\item{preprocessingServerInput}{reactives from preprocessingServer}

\item{qcServerInput}{reactives from qcServer}

\item{modelServerInput}{reactives from modelServer}

\item{inferenceServerInput}{reactives from inferenceServer}
}
\value{
nothing
}
\description{
Generates a single zip bundle containing the raw input file, annotation
file, the final QFeatures RDS (after modelling), an inference TSV, and
both report-preprocessing.Rmd and report-inference.Rmd filled with the
options selected in the app.
}
\keyword{internal}
