% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCombinedModule.R
\name{reportCombinedUI}
\alias{reportCombinedUI}
\title{Combined report UI}
\usage{
reportCombinedUI(id = "report")
}
\value{
A shiny tagList object that contains the combined report UI components
}
\description{
Combined report UI
}
\keyword{internal}
