% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDataProcessingModule.R
\name{reportDataProcessingServer}
\alias{reportDataProcessingServer}
\title{Server for data processing report}
\usage{
reportDataProcessingServer(
  id = "reportDataProcessing",
  variables,
  importServerInput,
  preprocessingServerInput
)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}

\item{importServerInput}{list with reactive values from importServer}

\item{preprocessingServerInput}{list with reactive values from preprocessingServer}
}
\value{
nothing
}
\description{
Server for data processing report
}
\keyword{internal}
