% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDataProcessingModule.R
\name{reportDataProcessingUI}
\alias{reportDataProcessingUI}
\title{Data processing report UI}
\usage{
reportDataProcessingUI(id = "reportDataProcessing")
}
\value{
A shiny tagList object that contains the data processing report UI components
}
\description{
Data processing report UI
}
\keyword{internal}
