% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPreprocessingModule.R
\name{reportPreprocessingServer}
\alias{reportPreprocessingServer}
\title{Server for preprocessing report}
\usage{
reportPreprocessingServer(
  id = "reportPreprocessing",
  variables,
  importServerInput,
  preprocessingServerInput
)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}

\item{importServerInput}{list with reactive values from importServer}

\item{preprocessingServerInput}{list with reactive values from preprocessingServer}
}
\value{
nothing
}
\description{
Server for preprocessing report
}
\keyword{internal}
