% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPreprocessingModule.R
\name{reportPreprocessingUI}
\alias{reportPreprocessingUI}
\title{Preprocessing report UI}
\usage{
reportPreprocessingUI(id = "reportPreprocessing")
}
\value{
A shiny tagList object that contains the preprocessing report UI components
}
\description{
Preprocessing report UI
}
\keyword{internal}
