% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModelingModule.R
\name{reportUI}
\alias{reportUI}
\title{report UI}
\usage{
reportUI(id = "report")
}
\value{
A shiny tagList object that contains the report UI components
}
\description{
report UI
}
\keyword{internal}
