% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMsqrob2App.R
\name{launchMsqrob2App}
\alias{launchMsqrob2App}
\title{Launches the msqrob2 Shiny App}
\usage{
launchMsqrob2App(maxSize = 500, launch.browser = TRUE)
}
\arguments{
\item{maxSize}{maximum memory size that input files are allowed to have in Mb}

\item{launch.browser}{logical; if TRUE (default) the app opens in the system browser}
}
\value{
shiny application object
}
\description{
Launches the msqrob2 Shiny App
}
\examples{
\dontrun{launchMsqrob2App()}

}
