% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMsqrob2ModelingApp.R
\name{launchMsqrob2ModelingApp}
\alias{launchMsqrob2ModelingApp}
\title{Launches the msqrob2 Shiny App}
\usage{
launchMsqrob2ModelingApp(maxSize = 500)
}
\arguments{
\item{maxSize}{maximum memory size that input files are allowed to have in Mb}
}
\value{
shiny application object
}
\description{
Launches the msqrob2 Shiny App
}
\examples{
\dontrun{launchMsqrob2ModelingApp()}

}
