% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchMsqrob2PreprocessingApp.R
\name{launchMsqrob2PreprocessingApp}
\alias{launchMsqrob2PreprocessingApp}
\title{Launches the msqrob2 Shiny App}
\usage{
launchMsqrob2PreprocessingApp(maxSize = 500)
}
\arguments{
\item{maxSize}{maximum memory size that input files are allowed to have in Mb}
}
\value{
shiny application object
}
\description{
Launches the msqrob2 Shiny App
}
