% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2DataProcessingServer.R
\name{msqrob2DataProcessingServer}
\alias{msqrob2DataProcessingServer}
\title{Shiny app server function for the data processing app}
\usage{
msqrob2DataProcessingServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function for the data processing app
}
