% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2ExploreServer.R
\name{msqrob2ExploreServer}
\alias{msqrob2ExploreServer}
\title{Shiny app server function for the msqrob2 Explore app}
\usage{
msqrob2ExploreServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function for the msqrob2 Explore app
}
