% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2ExploreUI.R
\docType{data}
\name{msqrob2ExploreUI}
\alias{msqrob2ExploreUI}
\title{Shiny app UI for the msqrob2 Explore app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
msqrob2ExploreUI
}
\description{
Shiny app UI for the msqrob2 Explore app
}
\keyword{datasets}
