% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2ModelingServer.R
\name{msqrob2ModelingServer}
\alias{msqrob2ModelingServer}
\title{Shiny app server function}
\usage{
msqrob2ModelingServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function
}
