% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2ModellingServer.R
\name{msqrob2ModellingServer}
\alias{msqrob2ModellingServer}
\title{Shiny app server function for the msqrob2 Modelling app}
\usage{
msqrob2ModellingServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function for the msqrob2 Modelling app
}
