% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2ModellingUI.R
\docType{data}
\name{msqrob2ModellingUI}
\alias{msqrob2ModellingUI}
\title{Shiny app UI for the msqrob2 Modelling app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
msqrob2ModellingUI
}
\description{
Shiny app UI for the msqrob2 Modelling app
}
\keyword{datasets}
