% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2PreprocessingServer.R
\name{msqrob2PreprocessingServer}
\alias{msqrob2PreprocessingServer}
\title{Shiny app server function}
\usage{
msqrob2PreprocessingServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function
}
