% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2PreprocessingUI.R
\docType{data}
\name{msqrob2PreprocessingUI}
\alias{msqrob2PreprocessingUI}
\title{Shiny app server object}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
msqrob2PreprocessingUI
}
\description{
Shiny app server object
}
\keyword{datasets}
