% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2UI.R, R/msqrob2ModelingUI.R
\docType{data}
\name{msqrob2UI}
\alias{msqrob2UI}
\title{Shiny app server object}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.

An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
msqrob2UI

msqrob2UI
}
\description{
Shiny app server object

Shiny app server object
}
\keyword{datasets}
