% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{NewPlotDensities}
\alias{NewPlotDensities}
\title{Plot intensity density curves per sample}
\usage{
NewPlotDensities(pe, assayName, varName, common = FALSE)
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to plot.}

\item{varName}{Character. Name of a \code{colData} variable used to colour
the curves.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot intensity density curves per sample
}
\keyword{internal}
