% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{PlotChargeStates}
\alias{PlotChargeStates}
\title{Stacked bar chart of precursor charge states per sample}
\usage{
PlotChargeStates(pe, assayName, chargeVar = "Precursor.Charge")
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to plot.}

\item{chargeVar}{Character. Name of the \code{rowData} column containing the
precursor charge state. Defaults to \code{"Precursor.Charge"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Only charge states 1–4 are shown. If the charge variable is not present in
\code{rowData}, a blank plot with an informative message is returned.
}
\keyword{internal}
