% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{PlotDimReduction}
\alias{PlotDimReduction}
\title{Plot precomputed dimensionality reduction}
\usage{
PlotDimReduction(dimRedResult, varName)
}
\arguments{
\item{dimRedResult}{List returned by \code{ComputeDimReduction}.}

\item{varName}{Character. Name of a \code{colData} variable for colouring.}
}
\value{
A \code{ggplot} object.
}
\description{
Renders a scatter plot from coordinates produced by
\code{ComputeDimReduction}, coloured by a selected \code{colData} variable.
Changing \code{varName} does not trigger recomputation.
}
\keyword{internal}
