% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{PlotIdentifications}
\alias{PlotIdentifications}
\title{Bar chart of identifications per sample}
\usage{
PlotIdentifications(pe, assayName, varName)
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to plot.}

\item{varName}{Character. Name of a \code{colData} variable used to colour
the bars.}
}
\value{
A \code{ggplot} object.
}
\description{
Counts the number of distinct features with a non-missing intensity value
for each sample, coloured by a selected \code{colData} variable.
}
\keyword{internal}
