% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{PlotNormBoxplots}
\alias{PlotNormBoxplots}
\title{Boxplot of sample intensities}
\usage{
PlotNormBoxplots(pe, assayName, varName, common = FALSE)
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to plot.}

\item{varName}{Character. Name of a \code{colData} variable used to colour
the boxes.}
}
\value{
A \code{ggplot} object.
}
\description{
Boxplot of sample intensities
}
\keyword{internal}
