% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{stepAggr}
\alias{stepAggr}
\title{Aggregate features to protein level}
\usage{
stepAggr(qf, aggrMethod, nameNormAssay, aggrCol, nameAggrAssay, nprecFilter)
}
\arguments{
\item{qf}{A \code{QFeatures} object.}

\item{aggrMethod}{Character. Aggregation method. One of \code{"maxLFQ"},
\code{"medianPolish"}, \code{"robustSummary"}, \code{"colMeans"},
\code{"colMedians"}, \code{"colSums"}.}

\item{nameNormAssay}{Character. Name of the normalised input assay.}

\item{aggrCol}{Character. Name of the \code{rowData} column used to group
features for aggregation (e.g. protein group column).}

\item{nameAggrAssay}{Character. Name given to the aggregated assay.}

\item{nprecFilter}{Integer. Minimum number of precursors required per
protein per sample; entries below this count are set to \code{NA}.}
}
\value{
A \code{QFeatures} object with the aggregated assay appended.
}
\description{
Aggregate features to protein level
}
\keyword{internal}
