% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{stepFilter}
\alias{stepFilter}
\title{Filter features based on a list of filter expressions}
\usage{
stepFilter(qf, filters)
}
\arguments{
\item{qf}{A \code{QFeatures} object.}

\item{filters}{A list of character strings, each a valid R filter expression
referencing columns of \code{rowData} (e.g. \code{"Decoy == 0"}).
All expressions are combined with \code{&}.}
}
\value{
A filtered \code{QFeatures} object.
}
\description{
Filter features based on a list of filter expressions
}
\keyword{internal}
