% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{stepLog}
\alias{stepLog}
\title{Log2-transform an assay}
\usage{
stepLog(qf, i, name)
}
\arguments{
\item{qf}{A \code{QFeatures} object.}

\item{i}{Character. Name of the assay to transform.}

\item{name}{Character. Name given to the resulting log-transformed assay.}
}
\value{
A \code{QFeatures} object with the log2-transformed assay appended.
}
\description{
Log2-transform an assay
}
\keyword{internal}
