% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{stepZeroToNA}
\alias{stepZeroToNA}
\title{Convert zero values to NA across all assays}
\usage{
stepZeroToNA(qf)
}
\arguments{
\item{qf}{A \code{QFeatures} object.}
}
\value{
A \code{QFeatures} object with zeros replaced by \code{NA}.
}
\description{
Convert zero values to NA across all assays
}
\keyword{internal}
