% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{utilitiesPreprocessing-imports
Histogram of proportion of missing values per feature}
\alias{utilitiesPreprocessing-imports
Histogram of proportion of missing values per feature}
\alias{PlotMissingValues}
\title{Internal utility imports}
\usage{
PlotMissingValues(pe, assayName, threshold)
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to plot.}

\item{threshold}{Numeric. The NA proportion threshold shown as a vertical
red dashed line.}
}
\value{
A \code{ggplot} object.
}
\description{
Internal utility imports
}
\keyword{internal}
