% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModule.R, R/qcModule.R
\name{inputServer}
\alias{inputServer}
\title{Server for input tab}
\usage{
inputServer(id = "import", variables)

inputServer(id = "import", variables)
}
\arguments{
\item{id}{module id}

\item{variables}{global reactive values object to share objects across modules}
}
\value{
list of reactive inputs

list of reactive inputs
}
\description{
Server for input tab

Server for input tab
}
\keyword{internal}
