% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModule.R, R/qcModule.R
\name{importUI}
\alias{importUI}
\alias{qcUI}
\title{input UI}
\usage{
importUI(id = "import")

qcUI(id = "qc", showFileInput = TRUE)
}
\value{
A shiny tagList object that contains the input UI components

A shiny tagList object that contains the input UI components
}
\description{
input UI

qc UI
}
\keyword{internal}
