% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2DataProcessingUI.R
\docType{data}
\name{msqrob2DataProcessingUI}
\alias{msqrob2DataProcessingUI}
\title{Shiny app UI object for the data processing app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
msqrob2DataProcessingUI
}
\description{
Shiny app UI object for the data processing app
}
\keyword{datasets}
