% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msqrob2UI.R
\docType{data}
\name{msqrob2UI}
\alias{msqrob2UI}
\title{Shiny app server object}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
msqrob2UI
}
\description{
Shiny app server object
}
\keyword{datasets}
