% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesPreprocessing.R
\name{PlotDimReduction}
\alias{PlotDimReduction}
\title{Dimensionality reduction plot}
\usage{
PlotDimReduction(pe, assayName, varName, method = c("MDS", "OmicsGMF"))
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{assayName}{Character. Name of the assay to use.}

\item{varName}{Character. Name of a \code{colData} variable used to colour
the points.}

\item{method}{Character. One of \code{"MDS"} or \code{"OmicsGMF"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Computes a 2D dimensionality reduction of samples and plots them coloured by
a selected \code{colData} variable. Supports MDS (via
\code{scater::runMDS}) and OmicsGMF (via \code{omicsGMF::runGMF}).
}
\keyword{internal}
